+++ packages/solidart/CHANGELOG.md
@@ -29,7 +29,7 @@
 
 ## 1.3.0
 
-- **FEAT**: Automatic disposal, [see the docs here](https://docs.page/nank1ro/solidart/advanced/automatic-disposal)
+- **FEAT**: Automatic disposal, [see the docs here](https://docs.page/nank1ro/solidart~85/advanced/automatic-disposal)
 
 ## 1.2.2